[mvp]: fix some error (#5)
* debug build config

* update aws server config

* debug workflow

* debug workflow

* debug workflow

* debug workflow

* add license & add openai api key

* feat(api): add task list api

* perf(task): avoid using S3 and output SQS within Lambda task functions

* perf(save_chunk_list): batch save chunks

* update lambda function requirements

* fix(lambda_handler): fix lambda_handler get_env_variable call error

* refactor(type): update pageparams type

* feat(api): add operation_id for api

* update makefile

* update ci-server.yml

* ci(server): update server install script

* fix actions

* update server-deploy

* update server-deploy

* update server-deploy

* update server-deploy

* fix(batch_execute_task): set batch_execute_task batch_size as 5

* fix(task): fix lambda task handler asyncio error

* ci(deploy): add env:TASK_ENGINE_CLASSNAME、DB_ENGINE_CLASSNAME

* ci(deploy): add api gateway

* update templage server yaml

* ci(deploy): use http replace rest

* update deploy config

* fix deploy parameter

* refactor(deploy): fix the id of aws whiskerHttpApi resource

* fix deploy parameter

* remove http api

* remove option method in AllowMethods

* fix template yaml

* fix template yaml

* fix template yaml

* add sqsQueueUrl

* debug sqs

* debug sqs

* update ParameterKey

* update task

* remove sleep when send message

* add response model

* feat: add cache for tenant verify

* update requirements.txt

* refactor(task): remove task pool

* refactor(task): update task deploy config

* fix build error

* fix build error

* update util file path

* fix(retriveal): fix retriveal timeout bug

* debug

* update util file path

* debug

* debug

* fix(knowledge): fix query with enum instance

* refactor(task): update lambda handler

* refactor(task): update lambda handler

* feat(task): delete knowledge chunks before insert chunks

* fix(task): fix restart_task request body

* fix(task): fix restart_task error

* refactor(dlq): remove dlq

* deploy preview

* refactor(plugin): dynamic plugin instantiation at runtime

* delete code commit

* feat(knowledge): can repeat add github repo now

* fix(knowledge): fix add knowledge bug

* fix(knowledge): delete repeat knowledge before add

* refactor(add_knowledge): add code commit

+++ server/plugins/db_engine/supabase/client.py
@@ -153,6 +153,11 @@ class SupaBasePlugin(DBPluginInterface):
     async def delete_knowledge(
         self, tenant_id: str, knowledge_id_list: List[str]
     ) -> List[Knowledge]:
+        if not knowledge_id_list:
+            return []
+        # delete task  and delete chunks
+        await self.delete_knowledge_task(tenant_id, knowledge_id_list)
+        await self.delete_knowledge_chunk(tenant_id, knowledge_id_list)
         res = (
             self.supabase_client.table(self.settings.KNOWLEDGE_TABLE_NAME)
             .delete()
@@ -164,6 +169,8 @@ class SupaBasePlugin(DBPluginInterface):
 
     # =============== chunk ===============
     async def save_chunk_list(self, chunk_list: List[Chunk]):
+        if not chunk_list:
+            return []
         res = (
             self.supabase_client.table(self.settings.CHUNK_TABLE_NAME)
             .insert(
@@ -191,6 +198,18 @@ class SupaBasePlugin(DBPluginInterface):
             self.supabase_client.table(self.settings.CHUNK_TABLE_NAME)
             .select("*")
             .eq("chunk_id", chunk_id)
+            .eq("tenant_id", tenant_id)
+            .execute()
+        )
+        return Chunk(**res.data[0]) if res.data else None
+
+    async def delete_knowledge_chunk(
+        self, tenant_id: str, knowledge_ids: str
+    ) -> List[Chunk] | None:
+        res = (
+            self.supabase_client.table(self.settings.CHUNK_TABLE_NAME)
+            .delete()
+            .in_("knowledge_id", knowledge_ids)
             .eq("tenant_id", tenant_id)
             .execute()
         )
@@ -240,6 +259,18 @@ class SupaBasePlugin(DBPluginInterface):
             .execute()
         )
         return Task(**res.data[0]) if res.data else None
+
+    async def delete_knowledge_task(
+        self, tenant_id: str, knowledge_ids: str
+    ) -> List[Task] | None:
+        res = (
+            self.supabase_client.table(self.settings.TASK_TABLE_NAME)
+            .delete()
+            .in_("knowledge_id", knowledge_ids)
+            .eq("tenant_id", tenant_id)
+            .execute()
+        )
+        return res
 
     # =============== tenant ===============
     async def save_tenant(self, tenant: Tenant) -> Tenant | None:
